+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### Added
 
+* ConfigDB Watcher can receive messages from the server via RabbitMQ.
 * Optional `directory_watcher.tolerations` and `directory_watcher.affinity` to `values.yaml`.
 * Test file pb-test-20260126-24294.tar.bz2 to Git LFS.